+++ dashboard/templatetags/custom_tags.py
@@ -22,6 +22,7 @@ from dashboard.constants import BRANCH_MAPPING_KEYS, TS_JOB_TYPES
 from dashboard.managers.graphs import GraphManager, ReportsManager
 from dashboard.managers.jobs import JobTemplateManager
 from dashboard.managers.packages import PackagesManager
+from dashboard.managers.inventory import ReleaseBranchManager
 
 
 register = template.Library()
@@ -29,6 +30,8 @@ register = template.Library()
 
 @register.filter
 def get_item(dict_object, key):
+    if not isinstance(dict_object, dict):
+        return ''
     return dict_object.get(key)
 
 
@@ -228,9 +231,13 @@ def tag_job_form(template_type):
         available_build_systems.append(relstream.relstream_built)
     if available_build_systems:
         return_value['build_systems'] = available_build_systems
-    packages = package_manager.get_package_name_tuple()
+    packages = package_manager.get_package_name_tuple(check_mapping=True)
     if packages:
         return_value['packages'] = packages
+    relbranch_manager = ReleaseBranchManager()
+    release_branches = relbranch_manager.get_relbranch_name_slug_tuple()
+    if release_branches:
+        return_value['releases'] = release_branches
     return return_value